parser: Initial inittab parser implementation
A very simple parser to read a colon separated inittab.

It reads lines into a 4096 chars buffer. If it finds line to be bigger
than that, halts with an error. Line reading is able to skip commentary
lines, those started with a '#' char. Note that *no* trimming is
performed now on lines.

Later, this buffer is broken into string tokens. Finally, strings are
properly converted to their final representation (be it a number, enum
or string) into a struct that holds inittab entries - `struct
inittab_entry`.

Some checking for the validity of tokens before their conversion is
done, so it prevents non-numeric or negative <order> and <core-id>, or
invalid values for <type>, but nothing else - like ensuring a unique
<safe-mode> entry.

A main file that simply reads and prints the content of an informed
inittab file is also provided, so one can see it working.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>