main: Start processes and handle their lifecycle
Processes all inittab entries and start them, in order. It waits some
time - defined via TIMEOUT_ONE_SHOT - for all process with a given order
to finish. However, it doesn't do anything now in case this timeout go
off - except print a message. This behaviour needs to be defined.

If init receives a SIGINT message, it starts process shutdown.
Initially, it will send SIGTERM to all running processes. It waits some
time for them to finish - defined via TIMEOUT_TERM. If any process is
still running after that time, it sends SIGKILL to them. Only after all
processes are done, it then starts, in order, the shutdown processes -
in the same way startup ones are processed.

Makefile was changed to add '-Wall', so we can be aware of some more
warnings.

Init shall complain if a inittab doesn't have the <safe-mode> entry, but
it doesn't use it now.

Requirement: 251161
Requirement: 251172

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>