tests: QEMU aided tests
init is the program that kernel calls to start userspace. As so, it can
be challenging to test it as a normal process on userspace.

To overcome these challenges, it's possible to use an emulator so init
run as the PID 1 process.

A script, `qemu-tests.sh` helps to automate these tests. They are
basically composed of `inittab` and `inspect` files. The first defines
what is to be run and the second inspect init log to see if everything
went as expected.

Check tests/README.tests for a deeper explanation of qemu test
mechanism.

NOTE: This commit does not add the expected kernel image and rootfs -
not sure how they will be produced yet.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>