general: Add rule to format code
`make format-code` will format code according to project standard.
Project standard tries to emulate that of Linux kernel. Is using
clang-format as the formating tool. A simple `.clang-format` was added
to define format. It's based on clang-format documentation example of
Linux kernel format.

Code was not re-formated at this point.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>