main: Identify and act when init is running inside a container
When running inside a container, init can't setup console or set as
section leader - because it already is.

This patch add a function to identify this case and act on both cases.

Note that other things - like disabling Ctrl+Alt+Del or dealing with
watchdog - also may fail to work, but init should already live well with
that.

Running inside a container is useful for testing purposes - allow, for
instance, init to be run under valgrind or with memory sanitisers.

Container specific code is ifdef'ed using NDEBUG macro - the same to
disable asserts. This way this code should not make its way to release
builds.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>