tests: Add failure tests
Tests using inittab that are expected to fail, in the sense that init
won't start, as their inittab files are invalid. Nevertheless, it is
expected that init will fail 'gracefully', not with a crash, so we test
for that.

When an inittab test file name starts with 'fail', as in `fail-1-inittab`,
it will be treated as expecting to fail. QEMU output will only be
verified to ensure that init did exit with EXIT_FAILURE.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>