lexer: Quoted tokens
Tokens may have quotes (both ' or ") that will set a range of characters
in which delimiter won't be matched. Once any quote character is found,
no more matches will be performed for delimiter until the same quote
character is found again.

For instance, using ' ' (space) as delimiter on following text:

    The quick "brown fox" 'jumps "over the" lazy' dog.

Will result in the following tokens:

    The
    quick
    "brown fox"
    'jumps "over the" lazy'
    dog.

Note that this patch doesn't remove the quotes. This is so that a text
like:

   NAME1="VALUE,NUMBER1",NAME2="VALUE,NUMBER2"

Using ',' as delimiter will yield to

    NAME1="VALUE,NUMBER1"
    NAME2="VALUE,NUMBER2"

tokens.

Note that user of lexer is able to define if lexing should take quotes
into account or not. If current buffer finishes before a quoted token is
closed, lexer will return TOKEN_UNFINISHED_QUOTE as result and output
string will point to unfinished token - so it can be used for logging,
for instance.

Current lexer user, inittab, changed to use new API without caring about
quotes.

This feature shall make sense to parse "fstab" files.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>