tests: Functional tests for fstab mount
Current functional inittab tests - that run on QEMU/container - moved to
`tests/data/qemu/inittab`, and new tests focusing on fstab added to
`tests/data/qemu/fstab`. Note that inittab directory has a
`default-fstab` and fstab directory has a `default-inittab`, so tests
can focus on one thing - leaving other fixed.

Note that fstab tests don't work on container. To have something to be
mounted, two new devices are present on QEMU, mapping `test-fs.ext4` and
`test-fs2.ext4` (the second one can be changed for another filesystem in
which we are interested) files. However, there's no simple way of doing
that for systemd-nspawn - usually involves some loop device trickery
that would have to be replicated for QEMU. As this is not being done
now, fstab tests are simply skipped on containers - valgrind and ASAN
testing.

More information on tests/README.tests.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>