main: Sanitize safe-mode shutdown
Safe-mode placeholder process is restarted every time it dies - but that
is not good if init is shutting down the system. This process should not
be terminated during shutdown - in fact, if a 'safe-shutdown' process
crashes, safe-mode should be called. This patch prevents init from
terminating safe-mode process - be it just the placeholder or the real
one.

An interesting question is: should safe-mode process be somehow informed
that system is going to shutdown? This is not addressed in this patch.
It will simply die on reboot system call, right now.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>