main: Start safe-mode when due
If a process tagged as safe on inittab crashes, safe-mode application
should start. As init already provisions a process for safe-mode, all
that it is necessary is to exec safe-mode application on placeholder
process.

To do so, pid 1 writes on pipe to placeholder process.

Finally, if placeholder process itself dies, it is restarted. But if
safe-mode application was already started and crashed, init simply exits
to panic the kernel. This should allow an external watchdog to take
control.

Requirement: 283064

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>