lexer: Add option to remove quotes from quoted tokens
Tokens may have delimiters enclosed in quotes, such delimiters are
ignored. However, currently, quotes make their way to the final token.
For instance, for string

    Alfa,"beta,gama",delta

using `,` as delimiter, following tokens will be returned:

    Alfa
    "beta,gama"
    delta

Note the quotes.

This patches adds a new parameter to `next_token`, `remove_quotes`,
that if true, will remove quotes of quoted tokens (it's ignored if
tokens are not quoted), so previous example would yield to

    Alfa
    beta,gama
    delta

tokens.

This should allow more flexibility in parsing quoted strings. Current
usage was updated to *not* use this feature. If in the future it's
deemed that they need quote removal, it may be the case to make this
behaviour the default.

Tests added to `tests/lexer_test.c`.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>