main: Not use bash anymore to exec new process
To ease parsing of command line, init was using bash as a wrapper on
`exec` call to (effectively) start a new process. Aiming to end this
runtime dependence, init now parses process command line.

cmdline.[c|h] adds this feature. It provides a method `parse_cmdline`
that returns a struct containing the environment variables and arguments
to be passed to `exec` function. As the parser ends up allocating some
memory, this memory should be freed using the `free_cmdline_contents`
function.

Currently, parsing is done on child process, before `exec` call. It would
be possible to perform parsing on pid 1, before `fork` call, but I saw
no advantage of any approach. This could be changed if the needs arise.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>