main: Only decrease pending process on STAGE_STARTUP and STAGE_SHUTDOWN
init keeps track of how many <one-shot> (during startup) and <shutdown>
(during shutdown) process are created with a given <order>, so it knows
when to start next round of process (with <order> bigger than current).

It increases this track for each new process created and decreases when
process finish. When counting reaches zero, init knows that it is time to
start next round of process.

However, when received signal to shut system down, init reset this
counter to zero as it enters STAGE_TERMINATION. In this stage, init will
send SIGTERM (and possibly SIGKILL) to currently running process.

But counter was still decreased, resulting in log messages 'Pending
decreased to -X' - obviously wrong for a counter.

After this stage, init moves to STAGE_SHUTDOWN, in which it will start
all process that need to run just before shutdown, in the order
specified on inittab. In this moment, init will again keep track of
processes.

This patch fixes counter decreasing, so that in only happens during
STAGE_STARTUP and STAGE_SHUTDOWN.

Tests also added to check for negative counters.

Fixes: 74c9482474fa1716bf825513e921a24802cc6e2f

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>