main: Prevent accidental closing of file descriptors
Methods `setup_stdio()` and `setup_stty()`, which run on children
process before `run_exec()` (and consequent change of process text) set
up children default file descriptors, `stdin`, `stderr` and `stdout`.

In this process, they use `dup2()` to set a preexistent file descriptor
to the default file descriptor - for instance, `setup_stdio()` dups log
fd on stderr and stdout. However, it may be possible that a preexistent
file descriptor has a number smaller or equal that of stderr. In this
case, `dup2()` will end up closing the preexistent file descriptor -
which may be undesirable, as it is when dups (and closes) log fd itself.

To prevent this, this patches introduces `safe_dup`, which must be
called to duplicate preexistent file descriptors to a number > stderr.
So, callers of `dup2()` can save any important file descriptor before
setting up default file descriptors.

Fixes: e01e5c55ff4bbbf4e175b3694055fd5c9b17be6b
Fixes: d69908a95df6d6bbebe7c61885a16643dd0bddd0

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>