main: Log improvements
Eventual calloc fail: It still to be decided what to do in this case
- of not being able to run `start_processes()` method, but let's at
least log it. It may appear on fault injection tests, it's useful to
be aware of that.

waitpid result: Right now, it's not expected `ECHILD` error to happen
on `waitpid` call, but fault injection library may inject it anyway.
So, let's log what happened after `waitpid` calls, so it's possible
to correctly interpret tests result.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>