tests: "Safe" executables are expected to handle SIGTERM
On slower environments, like QEMU without kvm, `/usr/bin/kill` command
to shut system down would not go before pid 1 sends SIGTERM to it,
effectively making it exit with exitcode != 0, which is unexpected for a
safe process (as defined on inittab).

This patch introduces `safe-kill` executable, a simple wrapper for `kill`
that ignores SIGTERM so it exits normally.

This patch also makes `safe-mode` application ignore SIGTERM - as a
safe application, it should be able to handle SIGTERM signals.

Reviewed-by: Flavio Ceolin <flavio.ceolin@intel.com>